Added eval function
Major updates to TODO, planning to revamp Array in major ways.
The original plan was to have Array behave much like a numpy array,
with a lot of runtime controls for things like row/col major ordering,
data ownership, etc. The flexibility offered by Expressions seems to
be a better alternative, and it will simplify a lot of the internals.
A new class 'View' will look a lot like the non-owning components of
Array, and will take the generic iterator with it.