Added broadcasting and square-bracket slicing
Can now 'broadcast' Dense objects, Numpy style. See DenseView.cpp
unit test for example. Some further work is needed to clean up the
syntax and make it a little less cumbersome. It may require a lot
of work on expressions.

At some point, it may be worth reviewing the SquBrktSlicer helper
class. It's necessary to add a little () at the end of slices to
'execute' the slice. Implicit conversion just didn't work. This
is a minor syntax concession, but it's not intuitive at all.