Improved striping interface further.
Introduced proxy object StripeGeneratorImpl which merely forwards
args to a StripeGenerator and provides a begin and end method.
The end method produces a std::size_t given by num_stripes, which
acts as a sentinel for the StripeGenerator. StripeGenerator no longer
decays to a bool, but can be compared to a std::size_t.