UNSTABLE STATE: Striped expressions
Overhaul of expressions to make use of striping under way.
A lot of cleanup left to do, and the future of cumulative
expressions looks bleak. Generator expressions may be salvageable,
but may need a substantial rework.
On the bright side, this is all very cool, and will open doors to
many much more useful functions.